2nd commit update directory refactoring and chatroom.html and main.html and review.html
