+++ src/main/java/kr/kro/namohagae/global/controller/GlobalController.java
@@ -13,7 +13,6 @@ import kr.kro.namohagae.board.service.*;
 import kr.kro.namohagae.global.security.MyUserDetails;
 import kr.kro.namohagae.global.service.ReportService;
 import kr.kro.namohagae.global.websocket.WebSocketService;
-import kr.kro.namohagae.mall.dto.AddressDto;
 import kr.kro.namohagae.mall.dto.ProductDto;
 import kr.kro.namohagae.mall.dto.QnaDto;
 import kr.kro.namohagae.mall.service.AddressService;
@@ -24,9 +23,9 @@ import kr.kro.namohagae.member.dto.DogDto;
 import kr.kro.namohagae.member.dto.MemberDto;
 import kr.kro.namohagae.member.service.DogService;
 import kr.kro.namohagae.member.service.MemberService;
-import kr.kro.namohagae.puchingtest.dto.ReviewDto;
-import kr.kro.namohagae.puchingtest.service.ChatService;
-import kr.kro.namohagae.puchingtest.service.ReviewService;
+import kr.kro.namohagae.puching.dto.ReviewDto;
+import kr.kro.namohagae.puching.service.ChatService;
+import kr.kro.namohagae.puching.service.ReviewService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.annotation.Secured;
