+++ src/main/java/kr/kro/namohagae/puching/controller/ReviewRestController.java
@@ -1,17 +1,13 @@
-package kr.kro.namohagae.puchingtest.controller;
+package kr.kro.namohagae.puching.controller;
 
 import kr.kro.namohagae.global.security.MyUserDetails;
 import kr.kro.namohagae.member.service.FollowService;
-import kr.kro.namohagae.puchingtest.dto.ReviewDto;
-import kr.kro.namohagae.puchingtest.service.ReviewService;
+import kr.kro.namohagae.puching.service.ReviewService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.ModelAndView;
-
-import java.security.Principal;
 
 @RestController
 @RequestMapping(value = "/api/v1")
