+++ src/main/java/kr/kro/namohagae/puching/service/PuchingService.java
@@ -1,15 +1,13 @@
-package kr.kro.namohagae.puchingtest.service;
+package kr.kro.namohagae.puching.service;
 
 import kr.kro.namohagae.member.dao.MemberDao;
-import kr.kro.namohagae.puchingtest.dao.ChatDao;
-import kr.kro.namohagae.puchingtest.dao.Puchingdao;
-import kr.kro.namohagae.puchingtest.dto.PuchingDto;
+import kr.kro.namohagae.puching.dao.ChatDao;
+import kr.kro.namohagae.puching.dao.Puchingdao;
+import kr.kro.namohagae.puching.dto.PuchingDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.io.IOException;
 import java.util.List;
-import java.util.NoSuchElementException;
 
 @Service
 public class PuchingService {
