+++ src/main/java/kr/kro/namohagae/puching/websocket/ChatWebSocketHandler.java
@@ -1,61 +1,60 @@
-package kr.kro.namohagae.puchingtest.websocket;
+package kr.kro.namohagae.puching.websocket;
 
-import kr.kro.namohagae.puchingtest.service.ChatService;
-import kr.kro.namohagae.puchingtest.service.PuchingService;
+
+import kr.kro.namohagae.puching.service.ChatService;
 import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
+
 import org.springframework.stereotype.Component;
 import org.springframework.web.socket.*;
-
 
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
 @Component
-public class ChatPuchingWebsocketHandler implements WebSocketHandler{
-
-    @Autowired
-    private ChatService chatService;
-    @Autowired
-    private PuchingService puchingService;
-
+public class ChatWebSocketHandler implements WebSocketHandler {
     private static final Map<String, WebSocketSession> sessions = new ConcurrentHashMap<>();
+
+    @Autowired
+    private ChatService service;
+
 
 
     @Override
     public void afterConnectionEstablished(WebSocketSession session) throws Exception {
+
         String username= session.getPrincipal().getName();
+        // WebSocketSession에서 사용자 정보를 저장합니다.
         session.getAttributes().put("username", username);
-        sessions.put(username,session);
+        // 저장된 사용자 정보를 출력합니다.
+        sessions.put(username, session);
+        // WebSocket 연결이 성공적으로 열리면 호출됩니다.
+
     }
 
     @Override
     public void handleMessage(WebSocketSession session, WebSocketMessage<?> message) throws Exception {
-
         String sendusername= session.getPrincipal().getName();
         String payload = message.getPayload().toString();
 
         JSONObject jsonPayload = new JSONObject(payload);
         String receiverUsername = jsonPayload.getString("receiverUsername");
         String messageContent = jsonPayload.getString("message");
+        String messageType= jsonPayload.getString("messageType");
 
         JSONObject responseJson = new JSONObject();
         responseJson.put("sendername", sendusername);
         responseJson.put("receivername", receiverUsername);
         responseJson.put("message", messageContent);
 
-        if(messageContent.equals("cancel")){
-            String content="<div class=\"puching-info\">\n" +
-                    "<span class=\"puching-title\">퍼칭이 취소 되었습니다</span>\n" +
-                    "</div>";
-        //퍼칭 서비스로 상대번호와 내번호로 신청상태인 퍼칭 테이블을 조회하고 있으면 취소상태로 바꾸고 메세지번호를 리턴
-           chatService.cancelPuchingMessage(sendusername,receiverUsername,content); //메세지 내용을 취소되었습니다 로 바꾸는거
-        }
 
-        if(messageContent.equals("accept")){
-            chatService.aceeptPuchingMessage(sendusername,receiverUsername);
-            //퍼칭 서비스로 상대번호와 내번호로 신청상태인 퍼칭테이블을 조회하고 있으면 수락상태로 바꾸고 퍼칭번호를 리턴 리뷰작성하러가기 페이지에는 모델에 퍼칭테이블 내용을 담아서쏜다
-            //서브쿼리
+        if(messageType.equals("text")){
+         Integer haveChatRoom= service.saveTextMessage(sendusername,receiverUsername,messageContent,"text"); //메세지 저장 실패하면 리턴값이
+            responseJson.put("haveChatRoom",haveChatRoom);
+        }
+        if(messageType.equals("image") || messageType.equals("puching")){
+           Integer haveChatRoom= service.existByChatRoom(sendusername,receiverUsername);
+            responseJson.put("haveChatRoom",haveChatRoom);
         }
 
 
@@ -71,24 +70,31 @@ public class ChatPuchingWebsocketHandler implements WebSocketHandler{
             // 수신자가 지정되지 않은 경우, 메시지를 전송하지 않습니다. 여기에 메세지전송 실패를 리턴해줘서 채팅창에 찍자
         }
 
+
+
     }
+
 
 
     @Override
     public void handleTransportError(WebSocketSession session, Throwable exception) throws Exception {
-        System.out.println("퍼칭웹소켓에러@@@@@@");
-        System.out.println(exception);
+        // WebSocket 통신 중 에러가 발생하면 호출됩니다.
+        System.out.println("웹소켓이 에러났당@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@");
+
     }
 
     @Override
-    public void afterConnectionClosed(WebSocketSession session, CloseStatus status) throws Exception {
-      String username= session.getPrincipal().getName();
+    public void afterConnectionClosed(WebSocketSession session, CloseStatus closeStatus) throws Exception {
+        String username= session.getPrincipal().getName();
         sessions.remove(username);
-        System.out.println(status);
+        System.out.println("웹소켓 닫혔다@@@@@@@@@@");
+
     }
 
     @Override
     public boolean supportsPartialMessages() {
-        return true;
+        return false;
     }
+
+
 }
