+++ src/main/resources/mybatis-mapper/puchingMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="kr.kro.namohagae.puchingtest.dao.Puchingdao">
-    <select id="findByUsers" resultType="kr.kro.namohagae.puchingtest.dto.PuchingDto$readUser">
+<mapper namespace="kr.kro.namohagae.puching.dao.Puchingdao">
+    <select id="findByUsers" resultType="kr.kro.namohagae.puching.dto.PuchingDto$readUser">
         SELECT * FROM (
         SELECT rownum rnum,m.* FROM
         (
