+++ src/main/resources/mybatis-mapper/reviewMapper.xml
@@ -3,8 +3,8 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="kr.kro.namohagae.puchingtest.dao.ReviewDao">
-    <select id="findContentByReceiverNo" resultType="kr.kro.namohagae.puchingtest.dto.ReviewDto$profile">
+<mapper namespace="kr.kro.namohagae.puching.dao.ReviewDao">
+    <select id="findContentByReceiverNo" resultType="kr.kro.namohagae.puching.dto.ReviewDto$profile">
         select * from
             (
                 select rownum as rnum, r.* from
@@ -17,7 +17,7 @@
             )
         where rnum&gt;=#{startRownum}
     </select>
-    <select id="findContentByWriterNo" resultType="kr.kro.namohagae.puchingtest.dto.ReviewDto$Information">
+    <select id="findContentByWriterNo" resultType="kr.kro.namohagae.puching.dto.ReviewDto$Information">
         select * from
             (
                 select rownum as rnum, r.* from
@@ -37,7 +37,7 @@
         select count(puching_review_no) from PUCHING_REVIEW where puching_review_writer = #{memberNo}
     </select>
 
-    <select id="reviewWriteInfo" resultType="kr.kro.namohagae.puchingtest.dto.ReviewDto$writeview">
+    <select id="reviewWriteInfo" resultType="kr.kro.namohagae.puching.dto.ReviewDto$writeview">
         select 'http://localhost:8081/api/v1/image/profile?name=' || sm.MEMBER_PROFILE_IMAGE as myImage,'http://localhost:8081/api/v1/image/profile?name=' || rm.MEMBER_PROFILE_IMAGE as receiverImage,sm.MEMBER_NICKNAME as myNickName,rm.MEMBER_NICKNAME as receiverNickName
              ,p.PUCHING_LATITUDE as latitude,p.PUCHING_LONGITUDE as longitude,p.PUCHING_PROMISE_DATE as promiseDate
         from MEMBER sm,MEMBER rm,PUCHING p
